Added k8s initilization, nodes joined

+++ prepare-k8s-account.sh
@@ -9,7 +9,10 @@ if [[ $? != "0" ]] ; then
   K8S_HOME=$( getent passwd k8s | cut -d: -f6 )
 fi
 K8S_HOME=$( getent passwd k8s | cut -d: -f6 )
-echo "export PATH=${PATH}:/usr/local/bin"
+grep -qF sr/local/bi ${K8S_HOME}/.bashrc
+if [[ $? != "0" ]]; then
+  echo "export PATH=${PATH}:/usr/local/bin" >>${K8S_HOME}/.bashrc
+fi
 
 if [[ ! -d ${K8S_HOME}/.ssh ]] ; then
   mkdir -p ${K8S_HOME}/.ssh || exit 1
@@ -27,8 +30,10 @@ chmod 700 ${K8S_HOME}/.ssh || exit 1
 chmod 600 ${K8S_HOME}/.ssh/k8s.key ${K8S_HOME}/.ssh/authorized_keys || exit 1
 
 if [[ ${HOSTNAME} == "proxy-lb" ]]; then
-  # we need this user to manage nginx (LB) service, for simplification, will be given root suto
-  echo "k8s  ALL=(ALL)       NOPASSWD: ALL" >/etc/sudoers.d/k8s-to-root
+  # we need this user to manage nginx (LB) service, for simplification, will be given root sudo
+  if [[ ! -f /etc/sudoers.d/k8s-to-root ]]; then
+    echo "k8s  ALL=(ALL)       NOPASSWD: ALL" >/etc/sudoers.d/k8s-to-root
+  fi
 fi
 
 echo App account k8s ready
